cache log analysis & fetch it from the frontend
The backend now saves a pre-parsed representation of the log to storage, and returns a SAS URI which the frontend uses to fetch it. This should significantly reduce bandwidth usage and response times, by avoiding the repeated parsing and double data transfer.

This implementation is messy, but it's an intermediate step towards later refactoring.